Añadiendo Escucha de una cancion y TOP de canciones mas escuchadas

+++ backend/src/main/java/com/redroosters/backend/dto/EscuchaResponseDTO.java
@@ -9,6 +9,6 @@ public record EscuchaResponseDTO(
         Long id,
         String nombreUsuario,
         String tituloCancion,
-        Integer vecesEscuchada,
+        Long vecesEscuchada,
         LocalDateTime ultimaEscucha
 ) {}
